move hooks to core
